rename warnings file to dusty_warnings
